Tactical Command/requirements.txt
weaviate-client>=3.25.0
openai>=1.0.0
tiktoken>=0.5.0